Optimize v4

+++ src/algorithms/fastod/attribute_pair.h
@@ -6,13 +6,13 @@ namespace algos::fastod {
 
 class AttributePair {
 private:
-    const std::pair<SingleAttributePredicate, int> pair_;
+    const std::pair<SingleAttributePredicate, size_t> pair_;
 
 public:
-    AttributePair(const SingleAttributePredicate& left, int right) noexcept;
+    AttributePair(const SingleAttributePredicate& left, size_t right) noexcept;
 
     const SingleAttributePredicate& GetLeft() const noexcept;
-    int GetRight() const noexcept;
+    size_t GetRight() const noexcept;
 
     std::string ToString() const;
 
@@ -26,8 +26,8 @@ namespace std {
 template <>
 struct hash<algos::fastod::AttributePair> {
     std::size_t operator()(const algos::fastod::AttributePair& pair) const {
-        auto left_hash = std::hash<int>()(pair.GetLeft().GetHashCode());
-        auto right_hash = std::hash<int>()(pair.GetRight());
+        auto left_hash = std::hash<size_t>()(pair.GetLeft().GetHashCode());
+        auto right_hash = std::hash<size_t>()(pair.GetRight());
         return left_hash ^ right_hash;
     }
 };
