+++ src/algorithms/fastod/predicates/single_attribute_predicate.h
@@ -10,24 +10,24 @@ namespace algos::fastod {
 
 class SingleAttributePredicate {
 private:
-    int attribute_;
+    size_t attribute_;
     Operator operator_;
     static std::vector<std::vector<SingleAttributePredicate>> cache_;
     
 public:
-    SingleAttributePredicate(int attribute, Operator const& op) noexcept;
+    SingleAttributePredicate(size_t attribute, Operator const& op) noexcept;
 
-    int GetAttribute() const noexcept;
+    size_t GetAttribute() const noexcept;
     Operator const& GetOperator() const noexcept;
 
     std::string ToString() const;
-    int GetHashCode() const noexcept;
+    size_t GetHashCode() const noexcept;
 
     bool Violate(DataFrame const& data,
-                 int first_tuple_index,
-                 int second_tuple_index) const noexcept;
+                 size_t first_tuple_index,
+                 size_t second_tuple_index) const noexcept;
 
-    static SingleAttributePredicate GetInstance(int attribute, Operator const& op);
+    static SingleAttributePredicate GetInstance(size_t attribute, Operator const& op);
 
     friend bool operator==(SingleAttributePredicate const& x, SingleAttributePredicate const& y);
 };
